Changed output assembly to x86_64

+++ ast.cpp
@@ -27,7 +27,7 @@ ProgramNode* ProgramNode::create()
 {
 	ProgramNode* node = new ProgramNode;
 	MemoryManager::addNode(node);
-	
+
 	return node;
 }
 
@@ -41,7 +41,7 @@ void ProgramNode::show(ostream& out, int depth) const
 	AstNode::show(out, depth);
 	for (list<AstNode*>::const_iterator i = children_.begin(); i != children_.end(); ++i)
 	{
-		(*i)->show(out, depth + 1);	
+		(*i)->show(out, depth + 1);
 	}
 }
 
@@ -49,7 +49,7 @@ LabelNode* LabelNode::create(const char* name)
 {
 	LabelNode* node = new LabelNode;
 	MemoryManager::addNode(node);
-	
+
 	node->name_ = name;
 	node->symbol_ = 0;
 	return node;
@@ -58,14 +58,14 @@ LabelNode* LabelNode::create(const char* name)
 void LabelNode::show(std::ostream& out, int depth) const
 {
 	indent(out, depth);
-	out << "Label: " << name_ << endl; 
+	out << "Label: " << name_ << endl;
 }
 
 NotNode* NotNode::create(ExpressionNode* expression)
 {
 	NotNode* node = new NotNode;
 	MemoryManager::addNode(node);
-	
+
 	node->expression_ = expression;
 	return node;
 }
@@ -80,7 +80,7 @@ BinaryOperatorNode* BinaryOperatorNode::create(ExpressionNode* lhs, Operator op,
 {
 	BinaryOperatorNode* node = new BinaryOperatorNode;
 	MemoryManager::addNode(node);
-	
+
 	node->lhs_ = lhs;
 	node->op_ = op;
 	node->rhs_ = rhs;
@@ -91,7 +91,7 @@ void BinaryOperatorNode::show(std::ostream& out, int depth) const
 {
 	AstNode::show(out, depth);
 	lhs_->show(out, depth + 1);
-	rhs_->show(out, depth + 1);	
+	rhs_->show(out, depth + 1);
 }
 
 const char* BinaryOperatorNode::str() const
@@ -104,7 +104,7 @@ const char* BinaryOperatorNode::str() const
 	case kDivide: return "BinaryOp(/)";
 	case kMod: return "BinaryOp(mod)";
 	}
-	
+
 	assert(false);
 }
 
@@ -112,7 +112,7 @@ LogicalNode* LogicalNode::create(ExpressionNode* lhs, Operator op, ExpressionNod
 {
 	LogicalNode* node = new LogicalNode;
 	MemoryManager::addNode(node);
-	
+
 	node->lhs_ = lhs;
 	node->op_ = op;
 	node->rhs_ = rhs;
@@ -123,7 +123,7 @@ void LogicalNode::show(std::ostream& out, int depth) const
 {
 	AstNode::show(out, depth);
 	lhs_->show(out, depth + 1);
-	rhs_->show(out, depth + 1);	
+	rhs_->show(out, depth + 1);
 }
 
 const char* LogicalNode::str() const
@@ -133,7 +133,7 @@ const char* LogicalNode::str() const
 	case kAnd: return "Logical(and)";
 	case kOr: return "Logical(or)";
 	}
-	
+
 	assert(false);
 }
 
@@ -141,7 +141,7 @@ ComparisonNode* ComparisonNode::create(ExpressionNode* lhs, ComparisonNode::Oper
 {
 	ComparisonNode* node = new ComparisonNode;
 	MemoryManager::addNode(node);
-	
+
 	node->lhs_ = lhs;
 	node->op_ = op;
 	node->rhs_ = rhs;
@@ -158,7 +158,7 @@ const char* ComparisonNode::str() const
 	case kGreaterOrEqual: return "Comparison(>=)";
 	case kLessOrEqual: return "Comparison(<=)";
 	}
-	
+
 	assert(false);
 }
 
@@ -166,14 +166,14 @@ void ComparisonNode::show(std::ostream& out, int depth) const
 {
 	AstNode::show(out, depth);
 	lhs_->show(out, depth + 1);
-	rhs_->show(out, depth + 1);	
+	rhs_->show(out, depth + 1);
 }
 
 VariableNode* VariableNode::create(const char* name)
 {
 	VariableNode* node = new VariableNode;
 	MemoryManager::addNode(node);
-	
+
 	node->name_ = name;
 	node->symbol_ = 0;
 	return node;
@@ -185,11 +185,11 @@ void VariableNode::show(std::ostream& out, int depth) const
 	out << "Variable: " << name_ << endl;
 }
 
-IntNode* IntNode::create(int value)
+IntNode* IntNode::create(long value)
 {
 	IntNode* node = new IntNode;
 	MemoryManager::addNode(node);
-	
+
 	node->value_ = value;
 	return node;
 }
@@ -204,7 +204,7 @@ BlockNode* BlockNode::create()
 {
 	BlockNode* node = new BlockNode;
 	MemoryManager::addNode(node);
-	
+
 	return node;
 }
 
@@ -218,7 +218,7 @@ void BlockNode::show(ostream& out, int depth) const
 	AstNode::show(out, depth);
 	for (list<StatementNode*>::const_iterator i = children_.begin(); i != children_.end(); ++i)
 	{
-		(*i)->show(out, depth + 1);	
+		(*i)->show(out, depth + 1);
 	}
 }
 
@@ -226,7 +226,7 @@ IfNode* IfNode::create(ExpressionNode* condition, StatementNode* body)
 {
 	IfNode* node = new IfNode;
 	MemoryManager::addNode(node);
-	
+
 	node->condition_ = condition;
 	node->body_ = body;
 	return node;
@@ -243,7 +243,7 @@ IfElseNode* IfElseNode::create(ExpressionNode* condition, StatementNode* body, S
 {
 	IfElseNode* node = new IfElseNode;
 	MemoryManager::addNode(node);
-	
+
 	node->condition_ = condition;
 	node->body_ = body;
 	node->else_body_ = else_body;
@@ -259,10 +259,10 @@ void IfElseNode::show(std::ostream& out, int depth) const
 }
 
 GotoNode* GotoNode::create(const char* target)
-{	
+{
 	GotoNode* node = new GotoNode;
 	MemoryManager::addNode(node);
-	
+
 	node->target_ = target;
 	return node;
 }
@@ -277,7 +277,7 @@ PrintNode* PrintNode::create(ExpressionNode* expression)
 {
 	PrintNode* node = new PrintNode;
 	MemoryManager::addNode(node);
-	
+
 	node->expression_ = expression;
 	return node;
 }
@@ -285,14 +285,14 @@ PrintNode* PrintNode::create(ExpressionNode* expression)
 void PrintNode::show(std::ostream& out, int depth) const
 {
 	AstNode::show(out, depth);
-	expression_->show(out, depth + 1);	
+	expression_->show(out, depth + 1);
 }
 
 ReadNode* ReadNode::create(const char* target)
 {
 	ReadNode* node = new ReadNode;
 	MemoryManager::addNode(node);
-	
+
 	node->target_ = target;
 	return node;
 }
@@ -307,7 +307,7 @@ WhileNode* WhileNode::create(ExpressionNode* condition, StatementNode* body)
 {
 	WhileNode* node = new WhileNode;
 	MemoryManager::addNode(node);
-	
+
 	node->condition_ = condition;
 	node->body_ = body;
 	return node;
@@ -324,7 +324,7 @@ AssignNode* AssignNode::create(const char* target, ExpressionNode* value)
 {
 	AssignNode* node = new AssignNode;
 	MemoryManager::addNode(node);
-	
+
 	node->target_ = target;
 	node->value_ = value;
 	return node;
@@ -333,9 +333,9 @@ AssignNode* AssignNode::create(const char* target, ExpressionNode* value)
 void AssignNode::show(std::ostream& out, int depth) const
 {
 	AstNode::show(out, depth);
-	
+
 	indent(out, depth + 1);
 	out << target_ << endl;
-	
-	value_->show(out, depth + 1);	
+
+	value_->show(out, depth + 1);
 }
